backend/requirements.txt
flask
flask-cors
scikit-learn
joblib
tldextract
numpy
pytest
requests
python-dotenv
xgboost
mysql-connector-python